refactor(fmt): start moving from efm to null-ls
There were some issues where my configuration for black and isort in efm
would not recognise the project `pyproject.toml` file.

Since efm was always tricky to configure anyway, it was a good time to
switch config to null-ls.

The remaining tools still on efm need to be tested before they are moved
over.